separate semester days from merged data
